优化客户相关，部分模板页面，模板通知相关

+++ EMT.DoneNOW.Core/SYS/sys_form_tmpl_opportunity.cs
@@ -17,7 +17,7 @@ namespace EMT.DoneNOW.Core
         [DataMember]
         public Int64 id { get; set; }
         [DataMember]
-        public Int32 form_tmpl_id { get; set; }
+        public Int64 form_tmpl_id { get; set; }
         [DataMember]
         public String opportunity_no { get; set; }
         [DataMember]
